move european monte_carlo into new class
