+++ options/monte_carlo/__init__.py
@@ -1,1 +1,1 @@
-from .monte_carlo_european import MonteCarlo
+from .monte_carlo_european import MonteCarloEuropean
